chore: 🤖 rebase
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-14401
