+++ ui/desktop/app/templates/scopes/scope/projects/settings/index.hbs
@@ -3,123 +3,5 @@
   SPDX-License-Identifier: BUSL-1.1
 }}
 
-<<<<<<< HEAD
-<SettingsCard @header={{t 'titles.application'}} @icon='boundary'>
-  <:body>
-    <div class='app-information'>
-      <div>
-        <Hds::Text::Display>
-          {{t 'titles.client'}}
-        </Hds::Text::Display>
-        <Hds::Text::Body @tag='p' @weight='medium'>
-          {{@model.formattedDesktopVersion}}
-        </Hds::Text::Body>
-      </div>
-      <div>
-        <Hds::Text::Display>
-          {{t 'titles.cli'}}
-        </Hds::Text::Display>
-        <Hds::Text::Body @tag='p' @weight='medium'>
-          {{@model.formattedCliVersion}}
-        </Hds::Text::Body>
-      </div>
-      <div>
-        <Hds::Text::Display>
-          {{t 'titles.cache-daemon'}}
-        </Hds::Text::Display>
-        <Hds::Text::Body @tag='p' @weight='medium'>
-          {{#if @model.formattedCacheVersion}}
-            {{@model.formattedCacheVersion}}
-            <Hds::Badge
-              @text={{t 'states.running'}}
-              @icon='running'
-              @color='success'
-            />
-          {{else}}
-            <Hds::Badge
-              @text={{t 'states.not-running'}}
-              @icon='stop-circle'
-              @color='neutral'
-            />
-          {{/if}}
-        </Hds::Text::Body>
-      </div>
-    </div>
-    {{#if @model.error}}
-      <Hds::Alert @type='inline' @color='warning' as |A|>
-        <A.Title>{{t 'errors.settings-cache.title'}}</A.Title>
-        {{#if @model.error.message}}
-          <A.Description>{{@model.error.message}}</A.Description>
-        {{/if}}
-      </Hds::Alert>
-    {{/if}}
-
-    <Hds::Form::Select::Field
-      name='theme'
-      @width='30%'
-      {{on 'change' this.toggleTheme}}
-      as |F|
-    >
-      <F.Label>{{t 'titles.color-theme'}}</F.Label>
-      <F.Options>
-        {{#each this.themes as |theme|}}
-          <option
-            value={{theme.value}}
-            selected={{eq this.session.data.theme theme.value}}
-          >
-            {{t (concat 'themes.' theme.label)}}
-          </option>
-        {{/each}}
-      </F.Options>
-    </Hds::Form::Select::Field>
-  </:body>
-</SettingsCard>
-
-<SettingsCard @header={{t 'resources.user.title'}} @icon='user'>
-  <:state>
-    {{#if this.isAuthenticated}}
-      <Hds::Badge
-        @text={{t 'states.autheticated'}}
-        @icon='check'
-        @color='success'
-      />
-    {{else}}
-      <Hds::Badge @text={{t 'states.unauthenticated'}} @color='neutral' />
-    {{/if}}
-  </:state>
-  <:body>
-    <div class='app-information'>
-      <div>
-        <Hds::Text::Display>
-          {{t 'resources.auth-method.title'}}
-        </Hds::Text::Display>
-        <Hds::Text::Body @tag='p' @weight='medium' data-test-auth-method>
-          {{this.authMethod}}
-        </Hds::Text::Body>
-      </div>
-      <div>
-        <Hds::Text::Display>
-          {{t 'form.username.label'}}
-        </Hds::Text::Display>
-        <Hds::Text::Body @tag='p' @weight='medium' data-test-username>
-          {{this.userInfo}}
-        </Hds::Text::Body>
-      </div>
-    </div>
-  </:body>
-
-  <:action>
-    <Hds::Button
-      data-test-signout-button
-      @text={{t 'actions.deauthenticate'}}
-      @color='critical'
-      @icon='sign-out'
-      @size='small'
-      {{on 'click' this.invalidateSession}}
-    />
-  </:action>
-</SettingsCard>
-=======
-<SettingsCard::Application @model={{@model}} />
+<SettingsCard::Application @model={{@model}} @toggle={{this.toggleTheme}} />
 <SettingsCard::User />
->>>>>>> b38a33704 (refactor: 💡 templates)
